+++ inc/question.class.php
@@ -813,7 +813,8 @@ class PluginFormcreatorQuestion extends CommonDBChild implements PluginFormcreat
             Supplier::class         => Supplier::getTypeName(2),
             Contact::class          => Contact::getTypeName(2),
             Contract::class         => Contract::getTypeName(2),
-            Document::class         => Document::getTypeName(2)],
+            Document::class         => Document::getTypeName(2),
+            Project::class          => Project::getTypeName(2)],
          __("Tools") => [
             Reminder::class         => __("Notes"),
             RSSFeed::class          => __("RSS feed")],
@@ -1319,6 +1320,12 @@ class PluginFormcreatorQuestion extends CommonDBChild implements PluginFormcreat
             $row['uuid']);
 
       $row['_skip_checks'] = true;
+
+      // escape text fields
+      foreach (['name', 'description'] as $key) {
+         $row[$key] = $DB->escape($row[$key]);
+      }
+
       $newQuestion_id = $newQuestion->add($row);
       if ($newQuestion_id === false) {
          return false;