+++ setup.php
@@ -33,9 +33,9 @@
 
 global $CFG_GLPI;
 // Version of the plugin
-define('PLUGIN_FORMCREATOR_VERSION', '2.7.0');
+define('PLUGIN_FORMCREATOR_VERSION', '2.8.0');
 // Schema version of this version
-define('PLUGIN_FORMCREATOR_SCHEMA_VERSION', '2.7');
+define('PLUGIN_FORMCREATOR_SCHEMA_VERSION', '2.8');
 // is or is not an official release of the plugin
 define('PLUGIN_FORMCREATOR_IS_OFFICIAL_RELEASE', true);
 
@@ -58,7 +58,7 @@ function plugin_version_formcreator() {
       return false;
    }
    $requirements = [
-      'name'           => _n('Form', 'Forms', 2, 'formcreator'),
+      'name'           => 'Form Creator',
       'version'        => PLUGIN_FORMCREATOR_VERSION,
       'author'         => '<a href="http://www.teclib.com">Teclib\'</a>',
       'homepage'       => 'https://github.com/pluginsGLPI/formcreator',