+++ tests/src/PluginFormcreatorTargetTicketDummy.php
@@ -39,8 +39,8 @@ class PluginFormcreatorTargetTicketDummy extends \PluginFormcreatorTargetTicket
       return \PluginFormcreatorTargetTicket::getTable();
    }
 
-   public function publicGetTargetEntity(\PluginFormcreatorFormAnswer $formanswer, $requesters_id) {
-      return $this->getTargetEntity($formanswer, $requesters_id);
+   public function publicSetTargetEntity($data, \PluginFormcreatorFormAnswer $formanswer, $requesters_id) {
+      return $this->SetTargetEntity($data, $formanswer, $requesters_id);
    }
 
    public function publicPrepareTemplate($template, \PluginFormcreatorFormAnswer $formAnswer, $disableRichText = false) {